refactor(scripts): move documentation tooling into scripts/docs/ (P4)
