+++ .github/workflows/documentation-garden.yml
@@ -75,11 +75,11 @@ jobs:
           DOCS_GARDEN_LANE: ${{ inputs.lane || 'auto' }}
           DOCS_GARDEN_SHARD: ${{ inputs.shard || '' }}
           DOCS_GARDEN_DRY_RUN: ${{ github.event_name == 'workflow_dispatch' && inputs.dry_run && 'true' || 'false' }}
-        run: node scripts/docs-garden-issue.mjs --json
+        run: node scripts/docs/docs-garden-issue.mjs --json
       - name: Synchronize the monthly navigation evaluation issue
         if: github.event_name == 'schedule'
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           DOCS_NAVIGATION_EVAL_REPO: ${{ github.repository }}
           DOCS_NAVIGATION_EVAL_DRY_RUN: "false"
-        run: node scripts/docs-navigation-eval.mjs --schedule-issue --json
+        run: node scripts/docs/docs-navigation-eval.mjs --schedule-issue --json
