+++ docs/adr/0040-bounded-documentation-garden-queue.md
@@ -105,9 +105,9 @@ Run the documentation garden through a deterministic, issue-only scheduler:
 ## Evidence
 
 - `.github/workflows/documentation-garden.yml`
-- `scripts/docs-garden-issue.mjs`
-- `scripts/docs-garden-issue-helpers.mjs`
-- `scripts/docs-garden-issue.test.mjs`
+- `scripts/docs/docs-garden-issue.mjs`
+- `scripts/docs/docs-garden-issue-helpers.mjs`
+- `scripts/docs/docs-garden-issue.test.mjs`
 - `scripts/lib/gh-issue-lifecycle.mjs` — the shared queue-state arbitration and
   marker parsing this fail-closed recovery depends on
 - `.agents/skills/doc-garden/SKILL.md`
