+++ docs/evals/documentation-navigation-fixtures.json
@@ -15,10 +15,11 @@
     "docs/evals/documentation-navigation-fixtures.json",
     "docs/evals/documentation-navigation-baseline-fixtures.json",
     "docs/evals/documentation-navigation-baseline.json",
-    "scripts/docs-navigation-eval.mjs",
-    "scripts/docs-navigation-eval-helpers.mjs",
-    "scripts/docs-navigation-eval-result.mjs",
-    "scripts/docs-navigation-eval.test.mjs"
+    "scripts/docs/docs-navigation-eval.mjs",
+    "scripts/docs/docs-navigation-eval-helpers.mjs",
+    "scripts/docs/docs-navigation-eval-result.mjs",
+    "scripts/docs/docs-navigation-eval-result-shape.mjs",
+    "scripts/docs/docs-navigation-eval.test.mjs"
   ],
   "questions": [
     {
