+++ package.json
@@ -27,12 +27,12 @@
     "agent:context-budget:test": "node scripts/context/agent-context-budget.test.mjs",
     "docs:index": "node scripts/context/docs-index.mjs",
     "docs:index:test": "node scripts/context/docs-index.test.mjs",
-    "docs:audit": "node scripts/docs-audit.mjs",
-    "docs:audit:test": "node scripts/docs-audit.test.mjs",
-    "docs:garden": "node scripts/docs-garden-issue.mjs",
-    "docs:garden:test": "node scripts/docs-garden-issue.test.mjs",
-    "docs:navigation-eval": "node scripts/docs-navigation-eval.mjs",
-    "docs:navigation-eval:test": "node scripts/docs-navigation-eval.test.mjs",
+    "docs:audit": "node scripts/docs/docs-audit.mjs",
+    "docs:audit:test": "node scripts/docs/docs-audit.test.mjs",
+    "docs:garden": "node scripts/docs/docs-garden-issue.mjs",
+    "docs:garden:test": "node scripts/docs/docs-garden-issue.test.mjs",
+    "docs:navigation-eval": "node scripts/docs/docs-navigation-eval.mjs",
+    "docs:navigation-eval:test": "node scripts/docs/docs-navigation-eval.test.mjs",
     "adr:check": "node scripts/check-adr-reminder.mjs",
     "adr:check:test": "node scripts/check-adr-reminder.test.mjs",
     "agent:autoreview": "./scripts/agent-autoreview.sh",
