+++ scripts/AGENTS.md
@@ -17,7 +17,7 @@ garden_lane: agent-entry-points
 ## Scope
 
 `scripts/` holds deploy wrappers, agent quality gates, code-health checks, and
-repo maintenance utilities. 174 files sit flat at the top level today.
+repo maintenance utilities. 164 files sit flat at the top level today.
 
 ## Target Layout
 
@@ -31,7 +31,7 @@ Files stay flat until their phase merges.
 | `workflows/`    | P1    | scripts backing Actions workflow jobs  |
 | `bootstrap/`    | P2    | container and hosted-session setup     |
 | `context/`      | P3    | agent context, budget, doc catalog     |
-| `docs/`         | P4    | catalog, audit, garden, nav eval       |
+| `docs/`         | P4    | audit planner, garden, navigation eval |
 | `pr/`           | P5    | PR and issue state projections         |
 | `supply-chain/` | P6    | lockfile, audit, pin, skew gates       |
 | `mcp/`          | P6    | MCP broker, launcher, config rendering |
@@ -40,7 +40,7 @@ Files stay flat until their phase merges.
 | `terraform/`    | P10   | movable Terraform guards and helpers   |
 | `gate/`         | P11   | quality-gate satellites                |
 
-Landed: P1, P2, P3, P6, P7. `lib/` (the shared tier) and
+Landed: P1, P2, P3, P4, P6, P7. `lib/` (the shared tier) and
 `production-infra-identity-contract/` predate the reorganization. `setup.sh`
 stays flat: `.config/wt.toml` runs that exact path as the Worktrunk pre-start
 hook, and eight docs name it.
@@ -59,7 +59,7 @@ pinned hashes, and identities stay with their domain.
 
 ## Why Files Stay Flat
 
-Seven mechanisms pin `scripts/` paths. A file one of them names moves only when
+Eight mechanisms pin `scripts/` paths. A file one of them names moves only when
 that mechanism moves with it, in the same PR.
 
 - **Autoreview runtime materialization.** `agent-autoreview.sh` names its
@@ -68,6 +68,12 @@ that mechanism moves with it, in the same PR.
 - **Gate source-directory guards.** `agent-quality-gate.sh` gates real-tree
   routing on `$script_source_dir == $repo_root/scripts`, leaving its stub-repo
   unit tests unaffected.
+- **Gate pre-push module pins.** `agent-quality-gate.sh` imports
+  `docs/docs-navigation-eval-helpers.mjs` from `$script_source_dir` to classify
+  routing-sensitive paths, and hashes that same literal in
+  `implementation_signature()`. Both run only at pre-push time. A stale path
+  loses routing and freezes the `--skip-if-fresh` stamp on every developer
+  machine while CI stays green, so the gate test asserts the literal resolves.
 - **Sentry suite manifest.** `sentry-suite-manifest.json` keys are exact
   repo-relative paths, reconciled against `findSentrySuites()` by set equality
   both ways. A moved or renamed suite fails the gate closed.
